+++ app/javascript/packs/lesson/utils/currentTabStates.js
