+++ app/javascript/packs/locales/index.js
@@ -1,5 +1,5 @@
-import en from './en';
-import ru from './ru';
+import en from './en.js';
+import ru from './ru.js';
 
 const resources = {
   en, ru,
