+++ app/views/web/languages/modules/lessons/show.html.slim
@@ -43,4 +43,4 @@
             = link_to t('.sign_in'), new_session_path, class: "btn btn-primary"
     - else
       .d-flex.flex-column.vh-100#basics-lesson-container
-      = javascript_pack_tag 'lessonContainer'
+      = javascript_pack_tag 'lesson'
